refactor get_ml_model and load_dataset into the ml_models and client folder
