+++ client/__init__.py
@@ -2,4 +2,4 @@
 # SPDX-License-Identifier: Apache-2.0
 
 from .os_ml_client_wrapper import OsMlClientWrapper
-from .helper import get_client_configs, get_client, get_index_size
+from .helper import get_client_configs, get_client, get_index_size, load_dataset
