+++ configs/__init__.py
@@ -2,16 +2,22 @@
 # SPDX-License-Identifier: Apache-2.0
 
 from .tasks import categories, tasks, PIPELINE_FIELD_MAP
-from .helper import get_config, validate_configs, parse_arg_from_configs
+from .helper import (
+    get_config,
+    validate_configs,
+    get_client_configs,
+    parse_arg_from_configs,
+    get_remote_connector_configs,
+)
 from .config import (
-    MINIMUM_OPENSEARCH_VERSION,
     ML_BASE_URI,
-    DELETE_RESOURCE_WAIT_TIME,
-    DELETE_RESOURCE_RETRY_TIME,
-    DEFAULT_LOCAL_MODEL_NAME,
-    DEFAULT_MODEL_VERSION,
-    DEFAULT_MODEL_FORMAT,
     DEFAULT_ENV_PATH,
     BASE_MAPPING_PATH,
+    DEFAULT_MODEL_FORMAT,
+    DEFAULT_MODEL_VERSION,
+    DEFAULT_LOCAL_MODEL_NAME,
+    DELETE_RESOURCE_WAIT_TIME,
+    DELETE_RESOURCE_RETRY_TIME,
     QANDA_FILE_READER_PATH,
+    MINIMUM_OPENSEARCH_VERSION,
 )
